Create linearSearchAlgorithm.cpp
Linear Search Algorithm is created. An element can be searched in an array by  giving the size of array as input, then storing all the elements(here declared as integer), then giving the element to be searched. It will output whether the element is present in the array or not.